pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mudenet-pytorch"
version = "0.1.0"
description = "Production-grade PyTorch reproduction of MuDeNet — multi-patch descriptor network for visual anomaly detection."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "mudenet-pytorch contributors"},
]
keywords = ["anomaly-detection", "pytorch", "mudenet", "industrial-inspection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch",
    "torchvision",
    "numpy",
    "scipy",
    "scikit-learn",
    "Pillow",
    "tqdm",
    "matplotlib",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
    "pre-commit",
]
viz = [
    "tensorboard",
]

[project.scripts]
mudenet = "mudenet.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # bugbear
    "SIM",   # simplify
    "RUF",   # ruff-specific
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["mudenet"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
    "torchvision.*",
    "scipy.*",
    "sklearn.*",
    "tqdm.*",
    "matplotlib.*",
    "yaml.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
